14._azure_functions/PythonHttpTriger/__init__.py
from fastapi import FastAPI, Request
import azure.functions as func


app = FastAPI()

@app.get("/hello/{name}")
async def get_name(name: str):
  return { "name": name }

def main(req: func.HttpRequest, context: func.Context) -> func.HttpResponse:
    return func.AsgiMiddleware(app).handle(req, context)
